industry-backend-main/industry-backend-main/requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
google-generativeai
python-dotenv